optimize code

+++ src/main/java/com/example/demo/service/impl/ProductServiceImpl.java
@@ -4,7 +4,7 @@ import com.example.demo.model.Product;
 import com.example.demo.repository.ProductRepository;
 import com.example.demo.service.ProductService;
 import org.springframework.stereotype.Service;
-
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,6 +16,10 @@ public class ProductServiceImpl implements ProductService {
     public ProductServiceImpl(ProductRepository productRepository) {
         this.productRepository = productRepository;
     }
+
+   // private SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+
+
 
     @Override
     public List<Product> productsWithThisRange(int lowerPrice, int upperPrice) {
@@ -33,8 +37,14 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public void deleteProductById(Long id) {
-        productRepository.deleteProductById(id);
+    public Boolean deleteProductById(Long id) {
+       Optional<Product> product= productRepository.findById(id);
+        if (product.isPresent()) {
+            productRepository.deleteProductById(id);
+            return true;
+        }
+        return  false;
+
     }
 
     @Override
@@ -44,6 +54,9 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public Product save(Product product) {
+        Date date = new Date();
+        product.setMCreatedDate(date);
+        product.setMActiveDate(date);
         return productRepository.save(product);
     }
